requirements.txt
anyio==4.6.2.post1
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.9
filelock==3.16.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.4
platformdirs==4.3.6
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.31.0
sniffio==1.3.1
urllib3==2.0.2
virtualenv==20.27.1
